Committing partial fix to merge the code changes from master to resolve the earlier fix issues.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -89,16 +89,27 @@ public class ImageController {
     //This method fetches the image with the corresponding id from the database and adds it to the model with the key as 'image'
     //The method then returns 'images/edit.html' file wherein you fill all the updated details of the image
 
+    //The method returns to the 'images/edit.html' page if the user is the owner of the image
+    //Else the method returns to the 'images/image.html' page with a message that only owner can edit the image
+
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        if (user.getId() == image.getUser().getId()) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            return showImage(imageId, image.getTitle(), model);
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -139,10 +150,24 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
-    @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+
+    //The method deletes the image and redirect to the 'images.html' page if the user is the owner of the image
+    //Else the method returns to the 'images/image.html' page with a message that only owner can delete the image
+
+    @RequestMapping(value = "/deleteImage", method = {RequestMethod.DELETE, RequestMethod.POST})
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+
+        if (user.getId() == image.getUser().getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            return showImage(imageId, image.getTitle(), model);
+        }
     }
